+++ environment.yml
@@ -1,0 +1,69 @@
+name: griml
+channels:
+  - defaults
+  - conda-forge
+dependencies:
+  - _libgcc_mutex=0.1
+  - _openmp_mutex=4.5
+  - ca-certificates=2022.3.29
+  - certifi=2021.10.8
+  - ld_impl_linux-64=2.35.1
+  - libffi=3.3
+  - libgcc-ng=9.3.0
+  - libgomp=9.3.0
+  - libstdcxx-ng=9.3.0
+  - ncurses=6.3
+  - openssl=1.1.1n
+  - pip=21.2.4
+  - python=3.8.13
+  - readline=8.1.2
+  - setuptools=61.2.0
+  - sqlite=3.38.2
+  - tk=8.6.11
+  - wheel=0.37.1
+  - xz=5.2.5
+  - zlib=1.2.12
+  - pip:
+    - affine==2.4.0
+    - attrs==21.4.0
+    - cachetools==5.0.0
+    - charset-normalizer==2.0.12
+    - click==8.1.2
+    - click-plugins==1.1.1
+    - cligj==0.7.2
+    - earthengine-api==0.1.306
+    - fiona==1.8.21
+    - future==0.18.2
+    - geopandas==0.10.2
+    - google-api-core==2.7.2
+    - google-api-python-client==1.12.11
+    - google-auth==2.6.6
+    - google-auth-httplib2==0.1.0
+    - google-cloud-core==2.3.0
+    - google-cloud-storage==2.3.0
+    - google-crc32c==1.3.0
+    - google-resumable-media==2.3.2
+    - googleapis-common-protos==1.56.0
+    - griml==0.0.2
+    - httplib2==0.20.4
+    - httplib2shim==0.0.3
+    - idna==3.3
+    - munch==2.5.0
+    - numpy==1.22.3
+    - pandas==1.4.2
+    - protobuf==3.20.1
+    - pyasn1==0.4.8
+    - pyasn1-modules==0.2.8
+    - pyparsing==3.0.8
+    - pyproj==3.3.1
+    - python-dateutil==2.8.2
+    - pytz==2022.1
+    - rasterio==1.3.9
+    - requests==2.27.1
+    - rsa==4.8
+    - scipy==1.8.0
+    - shapely==1.8.1.post1
+    - six==1.16.0
+    - snuggs==1.4.7
+    - uritemplate==3.0.1
+    - urllib3==1.26.9
