+++ src/griml/filter/__init__.py
@@ -1,0 +1,2 @@
+from griml.filter.filter_area import *
+from griml.filter.filter_margin import *
