+++ src/griml/merge/__init__.py
@@ -1,0 +1,1 @@
+from griml.merge.merge_vectors import *
