+++ src/griml/metadata/__init__.py
@@ -1,0 +1,4 @@
+from griml.metadata.assign_certainty import *
+from griml.metadata.assign_id import *
+from griml.metadata.assign_names import *
+from griml.metadata.assign_sources import *
